Moved TreeLayoutEngine static methods to TreeLayoutEngineHelpers
